+++ changelog.md
@@ -27,6 +27,7 @@ Saves from 4.x are not compatible with 5.0.
 
 * **[Campaign]** Naval control points will no longer claim ground objectives during campaign generation and prevent them from spawning.
 * **[Mission Generation]** Mission results and other files will now be opened with enforced utf-8 encoding to prevent an issue where destroyed ground units were untracked because of special characters in their names.
+* **[Mission Generation]** Fixed generation of landing waypoints so that the AI obeys them.
 * **[UI]** Selling of Units is now visible again in the UI dialog and shows the correct amount of sold units
 * **[UI]** Fixed bug where an incompatible campaign could be generated if no action is taken on the campaign selection screen.